requirements.txt
behave==1.2.6
Flask-RESTful==0.3.9
Flask-Testing==0.8.1
elasticsearch==7.13.1
docker-compose==1.29.2
